requirements.txt
python-dotenv
streamlit
python-docx
openai
PyMuPDF
dotenv
langchain